Replace the `u-wave-web` npm package by the precompiled version
Now `u-wave-web` is the compiled version of the web client. It's placed
in the `./npm` folder instead of in `packages/u-wave-web-middleware`.

The repo root becomes a non-publishable build package. Perhaps we could
publish it again if we have a usable system for modifications in the
future. Until then it is not useful to publish it to npm.